Refactor: Move `menu.config` package to `config` package for better organization and consistency.
