+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/config/MainConfig.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.menu.config;
+package com.ghostchu.quickshop.config;
 
 
 /*
