+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/menu/ShopBrowseMenu.java
@@ -20,7 +20,7 @@ package com.ghostchu.quickshop.menu;
 import com.ghostchu.quickshop.QuickShop;
 import com.ghostchu.quickshop.menu.browse.GroupedItemPage;
 import com.ghostchu.quickshop.menu.browse.ShopListPage;
-import com.ghostchu.quickshop.menu.config.GuiConfig;
+import com.ghostchu.quickshop.config.GuiConfig;
 import net.tnemc.menu.core.Menu;
 import net.tnemc.menu.core.Page;
