+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/menu/staff/PlayerSelectionPage.java
@@ -20,7 +20,7 @@ package com.ghostchu.quickshop.menu.staff;
 import com.ghostchu.quickshop.QuickShop;
 import com.ghostchu.quickshop.api.shop.Shop;
 import com.ghostchu.quickshop.api.shop.permission.BuiltInShopPermissionGroup;
-import com.ghostchu.quickshop.menu.config.GuiConfig;
+import com.ghostchu.quickshop.config.GuiConfig;
 import com.ghostchu.quickshop.menu.shared.ClearSearchAction;
 import com.ghostchu.quickshop.menu.shared.GuiChatAction;
 import net.tnemc.item.providers.SkullProfile;
